Rename msg router file name in subspaces
