+++ packages/desmos-std/src/subspaces/mod.rs
@@ -1,5 +1,5 @@
 pub mod models;
-pub mod msg_router;
+pub mod msg_builder;
 pub mod msg_routes;
 pub mod query_router;
 pub mod query_routes;
